requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
pyyaml>=6.0
gitpython>=3.1
python-dotenv>=1.0
fastapi>=0.115.0
uvicorn>=0.32.0
